USB: phy-msm-usb: Enable USB Interrupt only if ASYNC IRQ is not present
Some USB controllers have different IRQ line to exit from
LPM. This is ASYNC IRQ. If this IRQ is present, we dont
need to enable or disable the USB interrupt during otg
suspend and resume respectively.

There is a new requirement for XO shutdown during idle.
To allow XO in idle CPU states. When CPU is idle, if an
irq is enabled and is not capable of waking up from xo,
the xo shutdown is aborted. Hence non-wakeup capable irq
must be disabled.

Change-Id: I3197a1b413a1bf849db3572442cc8dbc5f264b53
CRs-Fixed: 520531
Signed-off-by: Sujeet Kumar <ksujeet@codeaurora.org>

Conflicts:
	drivers/usb/otg/msm_otg.c
	include/linux/usb/msm_hsusb.h

Change-Id: Ibe274843ab891be16cf8765fa89b2097ec389427